+++ vfscore/src/inode.rs
@@ -32,7 +32,7 @@ pub trait VfsInode: DowncastSync + VfsFile {
         _name: &str,
         _ty: VfsNodeType,
         _perm: VfsNodePerm,
-        _rdev: Option<u32>,
+        _rdev: Option<u64>,
     ) -> VfsResult<Arc<dyn VfsInode>> {
         Err(VfsError::NoSys)
     }
